+++ src/content/presenters/ConsoleFramePresenter.ts
@@ -1,0 +1,25 @@
+export default interface ConsoleFramePresenter {
+  initialize(): void;
+
+  blur(): void;
+
+  // eslint-disable-next-line semi
+}
+
+export class ConsoleFramePresenterImpl implements ConsoleFramePresenter {
+  initialize(): void {
+    let iframe = document.createElement('iframe');
+    iframe.src = browser.runtime.getURL('build/console.html');
+    iframe.id = 'vimvixen-console-frame';
+    iframe.className = 'vimvixen-console-frame';
+    document.body.append(iframe);
+  }
+
+  blur(): void {
+    let ele = document.getElementById('vimvixen-console-frame');
+    if (!ele) {
+      throw new Error('console frame not created');
+    }
+    ele.blur();
+  }
+}
